requirements.txt
djangorestframework
Django==3.0.3
psycopg2==2.8.4
Pillow==7.0.0
django-phonenumber-field==4.0.0
phonenumbers==8.11.5
django-crispy-forms==1.9.0
django-bootstrap4==1.1.1
